#2 Change: Gameからターンに関する処理をTurnManagerに分割
